+++ bindings/pydrake/solvers/BUILD.bazel
@@ -57,6 +57,7 @@ drake_pybind_library(
         "solvers_py_branch_and_bound.cc",
         "solvers_py_clp.cc",
         "solvers_py_csdp.cc",
+        "solvers_py_evaluators.cc",
         "solvers_py_gurobi.cc",
         "solvers_py_ipopt.cc",
         "solvers_py_mathematicalprogram.cc",
@@ -229,6 +230,14 @@ drake_py_unittest(
         ":solvers",
         "//bindings/pydrake:math_py",
         "//bindings/pydrake/common/test_utilities",
+        "//bindings/pydrake/common/test_utilities:scipy_stub_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "evaluators_test",
+    deps = [
+        ":solvers",
         "//bindings/pydrake/common/test_utilities:scipy_stub_py",
     ],
 )
